feat: Rename integration to Hassarr and update metadata
- Updated integration name in hacs.json and manifest.json.
- Created initial implementation of the Hassarr integration with support for Overseerr, Sonarr, and Radarr.
- Added configuration flow for user input and validation.
- Implemented service for requesting media with appropriate event handling.
- Established API communication with Overseerr, Radarr, and Sonarr.
- Added translation files for English and Spanish.
- Defined constants for configuration and service parameters.
