+++ custom_components/hassarr/config_flow.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
 
-from typing import Any, Dict, Optional
+from typing import Any, Dict
 import json
+import re
 import voluptuous as vol
 
 from homeassistant import config_entries
@@ -18,14 +19,18 @@ from .const import (
     CONF_PRESETS, CONF_DEFAULT_TV_SEASONS, DEFAULT_TV_SEASONS_CHOICES,
     CONF_OVERSEERR_SERVER_ID, CONF_OVERSEERR_PROFILE_ID_MOVIE, CONF_OVERSEERR_PROFILE_ID_TV,
 )
-from .api_overseerr import OverseerrClient
-from .api_arr import RadarrClient, SonarrClient
+from .api_common import OverseerrClient, RadarrClient, SonarrClient
 
 BACKEND_OPTIONS = ["overseerr", "arr"]
+URL_RE = re.compile(r"^https?://", re.I)
 
 
-class OverseerrConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
-    VERSION = 5
+def _valid_url(url: str) -> bool:
+    return bool(URL_RE.match(url.strip()))
+
+
+class HassarrConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
+    VERSION = 6
 
     async def async_step_user(self, user_input: Dict[str, Any] | None = None):
         errors: Dict[str, str] = {}
@@ -46,40 +51,42 @@ class OverseerrConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
             api_key = user_input[CONF_API_KEY].strip()
             default_tv = user_input[CONF_DEFAULT_TV_SEASONS]
 
-            session = async_get_clientsession(self.hass)
-            client = OverseerrClient(base_url, api_key, session)
-            if await client.ping():
-                # fetch services + profiles for step 2
-                radarr = await client.list_radarr()
-                sonarr = await client.list_sonarr()
-
-                movie_profiles = {}
-                tv_profiles = {}
-                try:
-                    default_radarr = next((s for s in radarr if s.get("isDefault")), radarr[0] if radarr else None)
-                    default_sonarr = next((s for s in sonarr if s.get("isDefault")), sonarr[0] if sonarr else None)
-                    if default_radarr:
-                        det = await client.get_radarr_details(default_radarr["id"])
-                        movie_profiles = {str(p["id"]): p["name"] for p in (det.get("profiles") or [])}
-                    if default_sonarr:
-                        det = await client.get_sonarr_details(default_sonarr["id"])
-                        tv_profiles = {str(p["id"]): p["name"] for p in (det.get("profiles") or [])}
-                except Exception:
-                    # If details fetch fails, keep empty choices
-                    pass
-
-                self._ovsr_ctx = {
-                    "base_url": base_url,
-                    "api_key": api_key,
-                    "default_tv": default_tv,
-                    "radarr": radarr,
-                    "sonarr": sonarr,
-                    "movie_profiles": movie_profiles,
-                    "tv_profiles": tv_profiles,
-                }
-                return await self.async_step_ovsr_selects()
-
-            errors["base"] = "cannot_connect"
+            if not _valid_url(base_url):
+                errors["base"] = "invalid_url"
+            else:
+                session = async_get_clientsession(self.hass)
+                client = OverseerrClient(base_url, api_key, session)
+                if await client.ping():
+                    try:
+                        radarr = await client.list_radarr()
+                        sonarr = await client.list_sonarr()
+                        movie_profiles = {}
+                        tv_profiles = {}
+                        try:
+                            default_radarr = next((s for s in radarr if s.get("isDefault")), radarr[0] if radarr else None)
+                            default_sonarr = next((s for s in sonarr if s.get("isDefault")), sonarr[0] if sonarr else None)
+                            if default_radarr:
+                                det = await client.get_radarr_details(default_radarr["id"])
+                                movie_profiles = {str(p["id"]): p["name"] for p in (det.get("profiles") or [])}
+                            if default_sonarr:
+                                det = await client.get_sonarr_details(default_sonarr["id"])
+                                tv_profiles = {str(p["id"]): p["name"] for p in (det.get("profiles") or [])}
+                        except Exception as e:  # noqa: BLE001
+                            # allow profile lists to be empty
+                            movie_profiles = movie_profiles or {}
+                            tv_profiles = tv_profiles or {}
+                        self._ovsr_ctx = {
+                            "base_url": base_url,
+                            "api_key": api_key,
+                            "default_tv": default_tv,
+                            "radarr": radarr,
+                            "sonarr": sonarr,
+                            "movie_profiles": movie_profiles,
+                            "tv_profiles": tv_profiles,
+                        }
+                        return await self.async_step_ovsr_selects()
+                else:
+                    errors["base"] = "cannot_connect"
 
         schema = vol.Schema({
             vol.Required(CONF_BASE_URL): str,
@@ -109,12 +116,11 @@ class OverseerrConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
             tp = user_input.get(CONF_OVERSEERR_PROFILE_ID_TV)
             if tp:
                 data[CONF_OVERSEERR_PROFILE_ID_TV] = int(tp)
-
-            await self.async_set_unique_id(f"overseerr:{ctx['base_url']}")
+            host_id = ctx['base_url'].split('://',1)[-1].rstrip('/')
+            await self.async_set_unique_id(f"overseerr:{host_id}")
             self._abort_if_unique_id_configured()
-            return self.async_create_entry(title="Media Requests (Overseerr)", data=data)
+            return self.async_create_entry(title="Hassarr (Overseerr)", data=data)
 
-        # Build dynamic choices
         radarr = ctx.get("radarr") or []
         sonarr = ctx.get("sonarr") or []
         server_choices = {}
@@ -151,26 +157,30 @@ class OverseerrConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
             sonarr_lang = user_input.get(CONF_SONARR_LANG_PROFILE)
             default_tv = user_input[CONF_DEFAULT_TV_SEASONS]
 
-            rc = RadarrClient(radarr_url, radarr_key, session)
-            sc = SonarrClient(sonarr_url, sonarr_key, session)
-            if await rc.ping() and await sc.ping():
-                await self.async_set_unique_id(f"arr:{radarr_url}|{sonarr_url}")
-                self._abort_if_unique_id_configured()
-                data = {
-                    CONF_BACKEND: "arr",
-                    CONF_RADARR_URL: radarr_url,
-                    CONF_RADARR_KEY: radarr_key,
-                    CONF_RADARR_ROOT: radarr_root,
-                    CONF_RADARR_PROFILE: radarr_prof,
-                    CONF_SONARR_URL: sonarr_url,
-                    CONF_SONARR_KEY: sonarr_key,
-                    CONF_SONARR_ROOT: sonarr_root,
-                    CONF_SONARR_PROFILE: sonarr_prof,
-                    CONF_SONARR_LANG_PROFILE: int(sonarr_lang) if sonarr_lang else None,
-                    CONF_DEFAULT_TV_SEASONS: default_tv,
-                }
-                return self.async_create_entry(title="Media Requests (Sonarr/Radarr)", data=data)
-            errors["base"] = "cannot_connect"
+            if not (_valid_url(radarr_url) and _valid_url(sonarr_url)):
+                errors["base"] = "invalid_url"
+            else:
+                rc = RadarrClient(radarr_url, radarr_key, session)
+                sc = SonarrClient(sonarr_url, sonarr_key, session)
+                if await rc.ping() and await sc.ping():
+                    host_id = f"{radarr_url.split('://',1)[-1]}|{sonarr_url.split('://',1)[-1]}".rstrip('/')
+                    await self.async_set_unique_id(f"arr:{host_id}")
+                    self._abort_if_unique_id_configured()
+                    data = {
+                        CONF_BACKEND: "arr",
+                        CONF_RADARR_URL: radarr_url,
+                        CONF_RADARR_KEY: radarr_key,
+                        CONF_RADARR_ROOT: radarr_root,
+                        CONF_RADARR_PROFILE: radarr_prof,
+                        CONF_SONARR_URL: sonarr_url,
+                        CONF_SONARR_KEY: sonarr_key,
+                        CONF_SONARR_ROOT: sonarr_root,
+                        CONF_SONARR_PROFILE: sonarr_prof,
+                        CONF_SONARR_LANG_PROFILE: int(sonarr_lang) if sonarr_lang else None,
+                        CONF_DEFAULT_TV_SEASONS: default_tv,
+                    }
+                    return self.async_create_entry(title="Hassarr (Sonarr/Radarr)", data=data)
+                errors["base"] = "cannot_connect"
 
         schema = vol.Schema({
             vol.Required(CONF_RADARR_URL): str,
@@ -188,7 +198,7 @@ class OverseerrConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
 
 
 @callback
-def async_get_options_flow(config_entry):
+def async_get_options_flow(config_entry):  # noqa: D401
     return OptionsFlowHandler(config_entry)
 
 
@@ -199,21 +209,18 @@ class OptionsFlowHandler(config_entries.OptionsFlow):
     async def async_step_init(self, user_input=None) -> FlowResult:
         errors = {}
 
-        # Common current values
         current_presets = self.entry.options.get(CONF_PRESETS, [])
         current_default = self.entry.options.get(
             CONF_DEFAULT_TV_SEASONS,
             self.entry.data.get(CONF_DEFAULT_TV_SEASONS, "season1"),
         )
 
-        # Try to fetch Overseerr servers/profiles if backend is Overseerr
-        ovsr_server_choices = {}
-        ovsr_movie_profiles = {}
-        ovsr_tv_profiles = {}
+        ovsr_server_choices: dict[str, str] = {}
+        ovsr_movie_profiles: dict[str, str] = {}
+        ovsr_tv_profiles: dict[str, str] = {}
         if self.entry.data.get(CONF_BACKEND) == "overseerr":
             try:
                 session = async_get_clientsession(self.hass)
-                from .api_overseerr import OverseerrClient
                 client = OverseerrClient(self.entry.data[CONF_BASE_URL], self.entry.data[CONF_API_KEY], session)
                 if await client.ping():
                     radarr = await client.list_radarr()
@@ -222,7 +229,6 @@ class OptionsFlowHandler(config_entries.OptionsFlow):
                         ovsr_server_choices[str(s["id"])] = f"Radarr: {s.get('name','Unnamed')} (#{s['id']})"
                     for s in sonarr:
                         ovsr_server_choices[str(s["id"])] = f"Sonarr: {s.get('name','Unnamed')} (#{s['id']})"
-
                     default_radarr = next((s for s in radarr if s.get("isDefault")), radarr[0] if radarr else None)
                     default_sonarr = next((s for s in sonarr if s.get("isDefault")), sonarr[0] if sonarr else None)
                     if default_radarr:
@@ -231,15 +237,13 @@ class OptionsFlowHandler(config_entries.OptionsFlow):
                     if default_sonarr:
                         det = await client.get_sonarr_details(default_sonarr["id"])
                         ovsr_tv_profiles = {str(p["id"]): p["name"] for p in (det.get("profiles") or [])}
-            except Exception:
+            except Exception:  # noqa: BLE001
                 pass
 
-        # Prepare defaults from existing options or data
         def get_opt_or_data(key):
             return self.entry.options.get(key, self.entry.data.get(key))
 
         if user_input is not None:
-            # Parse presets JSON first
             text = user_input["presets_json"]
             try:
                 data = json.loads(text)
@@ -252,12 +256,10 @@ class OptionsFlowHandler(config_entries.OptionsFlow):
                     if p["name"] in names:
                         raise ValueError(f"Duplicate preset name: {p['name']}")
                     names.add(p["name"])
-
                 out = {
                     CONF_PRESETS: data,
                     CONF_DEFAULT_TV_SEASONS: user_input[CONF_DEFAULT_TV_SEASONS],
                 }
-                # Optional Overseerr fields (only if backend is Overseerr)
                 if self.entry.data.get(CONF_BACKEND) == "overseerr":
                     sid = user_input.get(CONF_OVERSEERR_SERVER_ID)
                     if sid:
@@ -268,23 +270,19 @@ class OptionsFlowHandler(config_entries.OptionsFlow):
                     tp = user_input.get(CONF_OVERSEERR_PROFILE_ID_TV)
                     if tp:
                         out[CONF_OVERSEERR_PROFILE_ID_TV] = int(tp)
-
                 return self.async_create_entry(title="Options", data=out)
-            except Exception:
+            except Exception:  # noqa: BLE001
                 errors["base"] = "invalid_json"
 
-        # Build schema
         schema_dict = {
             vol.Required(CONF_DEFAULT_TV_SEASONS, default=current_default): vol.In(DEFAULT_TV_SEASONS_CHOICES),
             vol.Required("presets_json", default=json.dumps(current_presets, indent=2) if current_presets else "[]"): str,
         }
 
         if self.entry.data.get(CONF_BACKEND) == "overseerr":
-            # Use dropdowns if we have choices, else free-form numbers
             server_default = str(get_opt_or_data(CONF_OVERSEERR_SERVER_ID)) if get_opt_or_data(CONF_OVERSEERR_SERVER_ID) else None
             movie_prof_default = str(get_opt_or_data(CONF_OVERSEERR_PROFILE_ID_MOVIE)) if get_opt_or_data(CONF_OVERSEERR_PROFILE_ID_MOVIE) else None
             tv_prof_default = str(get_opt_or_data(CONF_OVERSEERR_PROFILE_ID_TV)) if get_opt_or_data(CONF_OVERSEERR_PROFILE_ID_TV) else None
-
             schema_dict[vol.Optional(CONF_OVERSEERR_SERVER_ID, default=server_default)] = (
                 vol.In(ovsr_server_choices) if ovsr_server_choices else vol.Coerce(int)
             )
