+++ custom_components/hassarr/manifest.json
@@ -1,0 +1,13 @@
+{
+  "domain": "hassarr",
+  "name": "Hassarr Media Requests (Overseerr / Sonarr+Radarr)",
+  "version": "0.6.0",
+  "config_flow": true,
+  "documentation": "https://github.com/Gangoke/Hassarr",
+  "issue_tracker": "https://github.com/Gangoke/Hassarr/issues",
+  "codeowners": [
+    "@Gangoke"
+  ],
+  "iot_class": "cloud_poll",
+  "requirements": []
+}
