--- custom_components/overseerr/manifest.json
@@ -1,13 +1,0 @@
-{
-  "domain": "overseerr",
-  "name": "Media Requests (Overseerr / Sonarr+Radarr)",
-  "version": "0.5.0",
-  "config_flow": true,
-  "documentation": "https://github.com/yourname/homeassistant-overseerr-requests",
-  "issue_tracker": "https://github.com/yourname/homeassistant-overseerr-requests/issues",
-  "codeowners": [
-    "@yourname"
-  ],
-  "iot_class": "cloud_push",
-  "requirements": []
-}
